Move `group_errors_by_kind` to the `common::report` module
The `alias` built-in reports each kind of operand error under its own
title rather than merging unrelated kinds under one misleading title.
Other built-ins that collect several operand errors need the same
grouping, so make the function public in the module that already holds
the merging utilities.

The function is generic over the error type instead of naming
`alias::semantics::Error`; its tests move along with it and use a local
enum in place of that type. No behavior changes.
