+++ yash-builtin/src/common/report.rs
@@ -19,6 +19,7 @@
 //! This module provides utilities for printing error messages and computing
 //! appropriate results (exit status and divert values) for built-ins.
 
+use std::mem::{Discriminant, discriminant};
 use std::ops::ControlFlow::{Break, Continue};
 use yash_env::Env;
 use yash_env::semantics::{Divert, ExitStatus};
@@ -266,6 +267,26 @@ where
     Some(first)
 }
 
+/// Groups the given errors by kind.
+///
+/// Errors of the same kind (that is, the same enum variant) end up in the same
+/// group regardless of where they occur among the operands, so that a built-in
+/// can report each kind as its own message rather than merging unrelated kinds
+/// of errors under one misleading shared title. The groups are returned in the
+/// order in which each kind first appears in `errors`.
+#[must_use]
+pub fn group_errors_by_kind<E>(errors: &[E]) -> Vec<(Discriminant<E>, Vec<&E>)> {
+    let mut groups: Vec<(_, Vec<_>)> = Vec::new();
+    for error in errors {
+        let kind = discriminant(error);
+        match groups.iter_mut().find(|(k, _)| *k == kind) {
+            Some((_, group)) => group.push(error),
+            None => groups.push((kind, vec![error])),
+        }
+    }
+    groups
+}
+
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -373,6 +394,47 @@ mod tests {
             SpanRole::Primary {
                 label: "label 3".into()
             }
+        );
+    }
+
+    /// Error enum used to exercise [`group_errors_by_kind`]
+    #[derive(Debug, Eq, PartialEq)]
+    enum TestError {
+        Foo(&'static str),
+        Bar(&'static str),
+    }
+
+    #[test]
+    fn group_errors_by_kind_merges_non_adjacent_errors_of_the_same_kind() {
+        let errors = [
+            TestError::Foo("a"),
+            TestError::Bar("b"),
+            TestError::Foo("c"),
+        ];
+
+        let groups = group_errors_by_kind(&errors);
+
+        assert_eq!(
+            groups,
+            [
+                (discriminant(&errors[0]), vec![&errors[0], &errors[2]]),
+                (discriminant(&errors[1]), vec![&errors[1]]),
+            ]
+        );
+    }
+
+    #[test]
+    fn group_errors_by_kind_orders_groups_by_first_occurrence() {
+        let errors = [TestError::Bar("b"), TestError::Foo("a")];
+
+        let groups = group_errors_by_kind(&errors);
+
+        assert_eq!(
+            groups,
+            [
+                (discriminant(&errors[0]), vec![&errors[0]]),
+                (discriminant(&errors[1]), vec![&errors[1]]),
+            ]
         );
     }
